Add CentOS support to AgoraKube
- The problem
Agorakube didn't have CentOS Support and it would be great to make it
work with CentOS and RHEL based distributions.

- Challenges
The main challenge in the CentOS Support was to configure yum repositories for different
components instead of apt.

The second challenge was the resolver as Ubuntu uses systemd resolver
which CentOS does not have, which is configured in kubelet config file.

- Other issues
CentOS does not have /usr/local/bin in the PATH by default and it does
not pick up the settings from /etc/profile in the single play-book
execution

- Solution
Move the repository configuration to separate task file and only include
the task file based on distribution

Add the PATH environment variable to all tasks using shell module.

Add the check for configuring systemd resolver and only use it if the
distribtuion is Ubuntu