Disable --profiling flag
Profiling allows for the identification of specific performance bottlenecks. It generates a 
significant amount of program data that could potentially be exploited to uncover system 
and program details. If you are not experiencing any bottlenecks and do not need the 
profiler for troubleshooting purposes, it is recommended to turn it off to reduce the 
potential attack surface.